// packages/sdkwork-agentstudio-pc-core/src/sdk/appSdkPort.ts
export interface PagePayload<T> {
  content?: T[];
  totalElements?: number | string;
  totalPages?: number | string;
  number?: number | string;
  size?: number | string;
  last?: boolean;
}
// ...
export interface OrderVO {
  orderId?: string;
  orderSn?: string;
  subject?: string;
  paidAmount?: number | string;
  totalAmount?: number | string;
  status?: string;
  statusName?: string;
  paymentProvider?: string;
  paymentMethod?: string;
  payTime?: string;
  createdAt?: string;
  updatedAt?: string;
}

export interface ProductVO {
  id?: string;
  title?: string;
  price?: number | string;
  sales?: number | string;
}

export type PageOrderVO = PagePayload<OrderVO>;
export type PageProductVO = PagePayload<ProductVO>;